[enumerateDevices] Extend the timeout delay for enumerateDevices
With crrev.com/c/1472675 we are adding processing in the enumerateDevices() code
path that will also probe audio paramters, not only video paramters.
This is done, however, asynchronously and requires longer time to be
satisfied depending on the audio devices to be probed. This causes the
following tests to fail.

This CL adjusts the timeouts to 5 seconds which should be enough.

BUG=chromium:731170
TEST=locally ran the tests on a CrOS device with Chromium built with crrev.com/c/1472675

Change-Id: I4389cda37b0e6d6c67ef8fcd33d9db2773bd6f66