video: Use longer timelimit for VDA tests
It usually takes more than 1.5 minutes to run tests that use Chrome's
video_decode_accelerator_unittests. And, the binary test retries
failed test cases multiple times in order to reduce flakiness.

So, if the test failed on some test cases, it exceeds the time limit.
As a result, it only reports that the test is killed due to timeout and
cannot show any failed test case names.

This CL makes the timelimit 4 minutes so that the test will be run
to the end even when it fails on several cases.

BUG=chromium:932348
TEST=run video.DecodeAccelH264 on daisy

Change-Id: I6c6e00baeb9ab442ad8ee6c4c31127a64af44a7b
Signed-off-by: Keiichi Watanabe <keiichiw@chromium.org>