video: Use longer timelimit for VDA tests
This CL doubles the time limit for video tests that run
video_decode_accelerator_unittest. This will fix their failures on
daisy due to timeout.

In addition, we will be able to get more informative messages when
the tests fail due to other reasons.
Currently, if the test failed on some test cases, the binary test
will retry them to reduce flakiness. But, this retries sometimes
cause timeout. In such case, the test is kileed and cannot report
failed test cases.
By this change, the test will be able to run to the end even if
there are some failed test cases.

BUG=chromium:932348
TEST=run video.DecodeAccelH264 on daisy

Change-Id: I6c6e00baeb9ab442ad8ee6c4c31127a64af44a7b
Signed-off-by: Keiichi Watanabe <keiichiw@chromium.org>