Update patch set 1
Patch Set 1:

> And we're sure we're just getting unlucky rather than having added some extraneous patch to the kernel that reduces ASLR (like https://bugs.chromium.org/p/chromium/issues/detail?id=582144)?

The thing is, it generally only flakes on one or two vmareas (out of dozens), and the one that causes the failure is not stable either. It also seems to happen across pretty much all boards so it wouldn't be tied to a specific kernel patch, unless we merged in minor revisions that contained something all at the same time.

I can look again but a broad-based failure this unreliable doesn't point to any root cause in particular...

> At the very least let's go up to 10 to see if flake goes away.

Ack.

Patch-set: 1
Reviewer: Gerrit User 1117055 <1117055@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>