tast: Don't backspace in VirtualKeyboardTyping.
VirtualKeyboardTyping presses a bunch of keys on the keyboard to
test that typing works. However, recently there's been flakiness
in the test, where pressing 2 backspaces seem to delete 3
characters.

We're not sure why this happens (maybe due to new backspace
feature where holding down backspace deletes a whole word).
For now, let's just remove backspace from the test so that we can
get it on the CQ. In the future, we would still like to include
backspace in the test.

TEST=veyron_minnie
BUG=chromium:879073

Change-Id: I49145239ee9a32ec53bcf162188871b88200f0a9