Run ui.Exceptions test in user session
I've been trying to restrict usage of sign-in profile in chrome, and
working on preventing browsers windows to be open in sign-in profile.
This would break ui.Exceptions test because the test is rtying to open a
tab in the default, sign-in profile.

To fix the discrepancy, change the test to do fake login and run the
test in the user session.

TEST=run the test with chrome checking a browser is not opened in the
sign-in profile.
BUG=chromium:916859

Change-Id: I3b8a60d661c2a578230f63ebc01f50a682bef891
Reviewed-on: https://chromium-review.googlesource.com/1482531
Commit-Ready: Toni Baržić <tbarzic@chromium.org>
Tested-by: Toni Baržić <tbarzic@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>