tast-tests: Update ui.ScreenLock test to use keyboard.
Update the ui.ScreenLock test to create a virtual keyboard
device so it can lock the screen using Search+L instead of
chrome.autotestPrivate.lockScreen, and also make it check
that the wrong password doesn't unlock the screen but that
the right password does.

This brings ui.ScreenLock to parity with the
desktopui_ScreenLocker Autotest test running on bvt-perbuild
(apart from not continuing to call the
autotestPrivate.lockScreen method, which doesn't seem useful
here).

BUG=chromium:930157
TEST=ran the test many times locally against a caroline DUT
     and didn't see any flakiness

Change-Id: Ibfd00168246186266da7f2cc557d6ed570a7ba6b
Reviewed-on: https://chromium-review.googlesource.com/1484796
Commit-Ready: Dan Erat <derat@chromium.org>
Tested-by: Dan Erat <derat@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>