arc.MIDIClient Add support for parsing delta-time values
The delta-time value that immediately precedes the MIDI note is a
variable-length quantity. This was causing the test to intermittently
fail, since it's length was earlier hardcoded.

This CL adds support to parse the variable-length time, and then extract
the MIDI note.

This CL also fixes a minor error printing issue, where the expected and
actual MIDI message byte data was being swapped while printing the error
message.

BUG=b:126215719
TEST=tast -verbose run -build=true pmalani-kevin arc.MIDIClient passes

Change-Id: I4f2bc9e0754e2be05a790e6b8df911363b9c109d