tast-tests: Crostini: close tabs opened by tests
The next test after LaunchBrowser takes screenshots of X11 and Wayland
apps started from the terminal; it seems that in some cases, the Chrome
window ends up on top of the app, so the screenshot doesn't contain the
expected app window.  To try to prevent this, close the browser tabs at
the end of each test before moving on to the screenshot tests.

Also explicitly close the "New Tab" tab opened by creating the Chrome
connection at the beginning of the test so the Chrome window won't be
left around once all the other tabs are closed, just to be sure it won't
overlap the x11 test app window.

BUG=chromium:935599
CQ-DEPEND=CL:1490031
TEST=tast run chromebook vm.CrostiniStartEverything

Change-Id: I43e5a482aebbdf521cd082025f4253c2f2431d4d
Signed-off-by: Daniel Verkamp <dverkamp@chromium.org>