tast-tests: Crostini: close window after browser test
The next test after LaunchBrowser takes screenshots of X11 and Wayland
apps started from the terminal; it seems that in some cases, the Chrome
window ends up on top of the app, so the screenshot doesn't contain the
expected app window.  To try to prevent this, close the browser window
at the end of the browser test before moving on to the screenshot tests.

BUG=chromium:935599
TEST=tast run chromebook vm.CrostiniStartEverything

Change-Id: I43e5a482aebbdf521cd082025f4253c2f2431d4d
Signed-off-by: Daniel Verkamp <dverkamp@chromium.org>