tast-tests: Add chrome.Conn.CloseTarget.
Add a new CloseTarget method to chrome.Conn that calls the
Chrome DevTools Protocol "CloseTarget" method in the Target
domain. This provides an easy way for tests to close tabs
that they've created.

Also update the chrome.LoggedIn precondition to use this
method instead of the chrome.windows JS API when closing
windows between tests. I'm hopeful that this will avoid
sporadic "websocket: bad handshake" errors that I've been
seeing, which I suspect may be caused by the cdp package
attempting to use a connection to a closed window.

Finally, add a new "locked" field to chrome.Conn and set it
on the object returned by chrome.Chrome.TestAPIConn to
prevent people from calling either Close or CloseTarget on
the connection to the test extension.

BUG=chromium:936072,chromium:925703
TEST=ran local_test_runner with tests that use
     chrome.LoggedIn in a loop and didn't see any errors
     when windows are closed between tests

Change-Id: Ia884eafba85ddc35562433f5e7ae65d2b51bf857