Update patch set 2
Patch Set 2: Verified+1

I updated this to use the new CloseTarget() API - it is much cleaner, and as a bonus, we don't have to worry about how to format JS in Go. :)

I did add an extra step to clean up the "New Tab" that is opened when the test starts - I believe this is opened when creating the Chrome connection with chrome.New().  I'm unsure if this is intended behavior or if we could change the chrome.New() code to avoid opening this tab at all.  I didn't want to change it myself since it is used by lots of other tests.

Patch-set: 2
Reviewer: Gerrit User 1297374 <1297374@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Verified=+1