platform.MemoryPressure: log wpr output
I never questioned my hyper-optimistic assumption
that we would not need the WPR log until WPR failed
mysteriously in the test lab---which happened immediately
upon deployment.  This CL fixes that, by redirecting WPR's
stdout and stderr to wpr.log.gz in the output directory,
and handling the various failure modes.

BUG=chromium:936559
TEST=ran test, observed log file content

Change-Id: I07d2205db2eb12b77a0f229a848aaf4fadb3e721