Update patch set 1
Patch Set 1:

I tested this with a complete test, but also by causing the test to crash in the middle (with a "stop ui" on the DUT, since control-C in the sdk prevents execution of the defers).  Worked both times.

Patch-set: 1
Reviewer: Gerrit User 1000028 <1000028@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>