Update patch set 1
Patch Set 1:

> Patch Set 1: -Code-Review
> 
> FYI - I am able to reproduce a similar issue on my kevin where wpr is dying before the test starts
> 
> I applied this CL and tried to use the logging but the log file seems truncated and doesn't contain anything useful:
> 
> zcat ./tests/platform.MemoryPressure/wpr.log.gz
> 
> gzip: ./tests/platform.MemoryPressure/wpr.log.gz: unexpected end of file
> 
> xxd   ./tests/platform.MemoryPressure/wpr.log.gz
> 00000000: 1f8b 0800 0000 0000 00ff      
> 
> The file is almost empty

I added a comment to the bug, but my problem was that the paths we were passing to wpr were not correct.  I'm not sure why this would work sometimes but not others -- we can discuss further in the bug.  However, I think to be useful, this CL needs to produce a readable log file in this type of situation where wpr exits very quickly.

Patch-set: 1
Reviewer: Gerrit User 1000153 <1000153@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0