Update patch set 2
Patch Set 2:

OK, looks like it's triggered by e.g. a mapping only appearing in exactly one map. I'm not sure why this should happen but I've observed it happening with /etc/ld.so.cache so I'll come up with some other workaround for this.

The original test wouldn't catch this: it looks like if the mapping appeared in the first run but none of the others, then it would never be marked as failing since none of the other maps had it, and if it appeared later, it would be ignored.

Patch-set: 2
Reviewer: Gerrit User 1117055 <1117055@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>