security.ASLR: don't fail when encountering singleton mapping
If a mapping was only seen once across all iterations, then it would
spuriously be marked as unchanging even though it was never seen with
the same address again. To fix this we keep track of the number of
times we have seen that start address and ensure we only fail if we
see it more than once.

BUG=b:1495935
TEST=run on nautilus, inject fake constant mapping and ensure failure

Change-Id: I60afdfc9d11b13a733f0ff3a11aedd800f65c81c