security.ASLR: don't fail when encountering singleton mapping
If a mapping was only seen once across all iterations, then it would
spuriously be marked as unchanging even though it was never seen with
the same address again. To fix this we keep track of the number of
times we have seen that start address and ensure we only fail if we
see it more than once.

BUG=b:1495935
TEST=run on nautilus, inject fake constant mapping and ensure failure

Change-Id: I60afdfc9d11b13a733f0ff3a11aedd800f65c81c
Reviewed-on: https://chromium-review.googlesource.com/1496094
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>