platform.MemoryPressure: report page fault rate at each step
This adds a kernelmeter instance which reports page fault
stats after adding each tab.  This is useful for correlating
tab switch times and page fault rates, although increases in
the CPU load are not due only to increased swapping, but
also possibly by increased background activity from the
growing number of tabs, some of which may be still loading
in the background.

It is wasteful to have multiple sampling goroutines.  Ideally
the two meters would share one sampler.  However, the sampling
frequency is so low that the performance change doesn't justify
the extra work.

BUG=chromium:934937
TEST=ran the test

Change-Id: Ic762684fb55cdaf21666ac2a6ef32074e6f00159