tast: Add test for Japanese input on virtual keyboard.
We would like to have sanity checks for all supported languages on
Chrome OS. We'll start with Japanese, which has two different input
methods.

The test just loops through the two input methods and checks whether
the right suggestions show up when tapping certain keys.

We'll be adding additional languages in the future.

BUG=chromium:879073
TEST=veyron_minnie
Change-Id: I4bda68acc559c56fcc99a8c93845e77cc9614670