tast: Handle keys with same aria-label in VK tests.
In some cases, multiple keys may have the same aria-label in VK
tests. This is because the keyboard DOM may sometimes have
multiple "keyboards" (e.g. there is one keyboard per language),
but only one of the keyboards is active (the rest are hidden with
display: none).

To handle this, when we are tapping on keys, we'll now only tap
on client bounds that are valid, so that we tap on the active
keyboard instead of any hidden ones.

TEST=veyron_minnie
BUG=chromium:879073

Change-Id: Ica69b5430c7f544669b332d6b2725e137e7d4689
Reviewed-on: https://chromium-review.googlesource.com/1496186
Commit-Ready: Darren Shen <shend@chromium.org>
Tested-by: Darren Shen <shend@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>