tast-tests: CrostiniStartEverything: close chromebook tab
After CL:1489496, some test machines now open the Chromebook landing
page (https://www.google.com/chromebook/) in a new tab when the browser
tests open a fresh browser window.  This keeps the window open,
defeating the intent of the previous change (to clear the screen of
browser windows before the screenshot tests).  To remedy this,
explicitly search for and close the Chromebook tab if it exists.

I could not figure out how exactly this tab is launched; it would be
nice to have a cleaner way to prevent it from opening at all, if
possible, but this change is at least sufficient to fix the tests.

BUG=chromium:935599
TEST=Run vm.CrostiniStartEverything on banon

Change-Id: Id7aeb7c521a65b3f0a924ed350f0df90e8a75605
Signed-off-by: Daniel Verkamp <dverkamp@chromium.org>