platform.MemoryPressure: log zram stats and preallocate less
The current setting only allow the creation of 7 tabs on a
4GB kevin, which is too few, especially because more than 1 tab
can be discarded, thus affecting the initial tab set and causing
the test to fail.

Also, this problem may have occurred because of an unexpected
compression ratio.  With this change, the test also logs zram
stats.

BUG=chromium:934937
TEST=ran on 4GB minnie and others

Change-Id: I62cfd5485f7b09714825434d703551ee2bb2ce40
Reviewed-on: https://chromium-review.googlesource.com/1504107
Commit-Ready: Luigi Semenzato <semenzato@chromium.org>
Tested-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>