tast-tests: Crostini: close windows in LaunchTerminal
Now that the Chrome browser window has been closed before the screenshot
test, some runs are showing the Terminal window with vim on top of the
x11 demo app instead.  Close each Terminal window after it has been
verified to open in the LaunchTerminal subtest to clean up.

This is starting to feel like whack-a-mole, but I don't have a better
solution (adding delays between tests seems like it would be worse).

BUG=chromium:935599
TEST=tast run vm.CrostiniStartEverything on nami

Change-Id: I6f8cc0c87b59412aa1c7c87c468642caff052a2d
Signed-off-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1506174
Reviewed-by: Dan Erat <derat@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>