Add CCA API tast tests
This CL adds tast tests for verifying whether the private JavaScript APIs
CCA relies on work as expected. The APIs under testing are not owned by
CCA team. The test is used to prevent the implementation changes of those
APIs silently break CCA. It can also be used in auto-bisecting to find
out the culprit CL in a breakage.

BUG=b:117817410
TEST=Verified on Eve that the test passes.

Change-Id: Ifa65240cbd1c8cc4ced751218bafce5b3f118d48