platform.MemoryPressure: collect more vmstats
This extends the collection of page fault rates to any
of the counts in /proc/vmstats (limited to the ones of
immediate use: page faults, swap-ins, swap-outs, and
OOM kills).  Counts that aren't available in /proc/vmstat
in earlier kernel versions are treated as 0.

The data is exported by kernelmeter as VMStatsData.  This
change does not change the test functionality or output.
A following change will use the newly available data.

BUG=chromium:934937
TEST=ran the test

Change-Id: I15ef2fd76da23f6b7048b46096c21280714f3926