platform.MemoryPressure: measure swap during tab switches
We measure and report swap-in and swap-out rates and counts during
back-to-back tab switches on the initial tab set after opening
each new tab.

We also report zram stats, mainly as a sanity check (i.e. are
the compression rate and zram utilization reasonable).

This also fixes an error in the kernelmeter, which was reporting
absolute counts, instead of the deltas between reset time and
now.  (Those values were not used before this change so no
harm done.)

Finally, this adds OOM count reporting at the end of phases 1
and 2.

BUG=chromium:934937
TEST=ran test, checked numbers

Change-Id: Id97ce912ee68cf57796669f8b5c85e9d71bf2912
Reviewed-on: https://chromium-review.googlesource.com/1518413
Commit-Ready: Luigi Semenzato <semenzato@chromium.org>
Tested-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>