platform.MemoryPressure: measure swap during tab switches
We measure and report swap-in and swap-out rates and counts during
back-to-back tab switches on the initial tab set after opening
each new tab.

We also report zram stats, mainly as a sanity check (i.e. are
the compression rate and zram utilization reasonable).

The kernelmeter class adds a parameter to its constructor
that specifies whether max rate sampling is requested or not.

This also fixes an error in the kernelmeter, which was reporting
absolute counts, instead of the deltas between reset time and
now.  (Those values were not used before this change so no
harm done.)

BUG=chromium:934937
TEST=ran test, checked numbers

Change-Id: Id97ce912ee68cf57796669f8b5c85e9d71bf2912