Update patch set 2
Patch Set 2:

I understand this point and in fact I had considered it when I wrote the code.  However I doubt that the transformation is useful and I am concerned that it may be "premature abstraction".  The additional type introduces overhead (the struct declaration and an additional type name) but I never need to treat instances of that type separately from their array.  So in the places where now I write

  ... field1Array[i] ...
  ... field2Array[i] ...

I would rewrite those as

  ... array[i].field1 ...
  ... array[i].field2 ...

which is hardly an improvement.  You might argue that I should be writing

  element := &array[i]
  ... element.field1 ...
  ... element.field2 ...

but that's also isn't much of an improvement in performance, readability, ease of modification etc. even if I had more than 3 fields. I would agree with this change if I were using those arrays in more complicated ways, and I will make the transformation when and if that happens---like for instance I am doing now as I move from reading a single field of /proc/vmstat to multiple fields.

Sorry for pushing back, but I would very much appreciate an explanation for the perceived need for this transformation at this point, at the level of my discussion.

Patch-set: 2
Reviewer: Gerrit User 1000028 <1000028@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0