Update patch set 7
Patch Set 7:

(1 comment)

Please ping me if you need approvals, but I'm treating this as just FYI and it seems to all be shaping up fine. Definitely learning a lot from the Go style discussions.

One piece just to chime in on, to avoid having to do extra work

Patch-set: 7