tast-tests: Make security.Minijail not use [ -w /bin/bash ].
The "caps" test case in the security.Minijail test is
failing constistently in the lab because it expects [ -w
"$0" ] to succeed (where $0 is /bin/bash). This is
presumably failing due to the root partition being mounted
read-only.

I had ported this check from test-caps in the
security_Minijail0 Autotest test, and I'm not sure why it
would be working in that case. I'm changing the Tast test to
instead check /usr/local/bin, which seems like a safer
choice.

BUG=chromium:911879
TEST=ran it on a lab DUT (in addition to a local DUT) and
     verified that the "caps" test case now passes

Change-Id: I1f0b1fa0c87c39054ef8c93963815bf572b1cbe6