Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> (2 comments)
> 
> Added a HTTP timeout based on ctx.Duration. Trying to figure out what's next for this CL. Do you think this would work with the hardcoded "localhost:9999" URL? Or should I parameterize that before this CL can be landed?

it's probably preferable to parameterize that first. i suspect that this will fail in the lab in most (all?) cases otherwise, since servod typically won't be running at localhost:9999 on the system that's executing the test.

for the overall structure, i'm guessing that something like a Servo struct with an associated constructor would make sense:

  func New(ctx context.Context, host string, port int) (*Servo, error)

maybe we'd also want a function that automatically derives the connection info:

  func Default(ctx) (*Servo, error)

Default would need some way to find the appropriate servod instance when running in the lab, while falling back to localhost:9999 otherwise (assuming that that's the default location when a developer runs servod locally).

i might've mentioned it earlier, but this feels similar to the GAIA credentials use case at https://crbug.com/903521. when tast is running in the lab, we need a way to pass the servod info (which we can hopefully get from the tast.py autotest test that wraps tast) to tests. let's talk about this at the meeting today. probably we should start a doc. :-)

another thing to think about is how we want to prevent tests that require servo from running on devices that don't have one attached. i think i've heard that there's a goal for all lab devices to have servos connected to them, but i don't know if that's currently true. http://doc/1iYbKgW2syEguKfV6bCWFtmpWncpJsFZIGRm7vDICyQY touches on this.

Patch-set: 8
Reviewer: Gerrit User 1000042 <1000042@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>