Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> > Patch Set 8:
> > 
> > (2 comments)
> > 
> > Added a HTTP timeout based on ctx.Duration. Trying to figure out what's next for this CL. Do you think this would work with the hardcoded "localhost:9999" URL? Or should I parameterize that before this CL can be landed?
> 
> it's probably preferable to parameterize that first. i suspect that this will fail in the lab in most (all?) cases otherwise, since servod typically won't be running at localhost:9999 on the system that's executing the test.
> 
> for the overall structure, i'm guessing that something like a Servo struct with an associated constructor would make sense:
> 
>   func New(ctx context.Context, host string, port int) (*Servo, error)
> 
> maybe we'd also want a function that automatically derives the connection info:
> 
>   func Default(ctx) (*Servo, error)

Updated to add a Servo struct and a New function. Unsure how the connection info would be derived by Default so leaving that commented out for now. Also is it kosher to store ctx on Servo.ctx, or should ctx be passed into each method? (see newest patch)

> another thing to think about is how we want to prevent tests that require servo from running on devices that don't have one attached. i think i've heard that there's a goal for all lab devices to have servos connected to them, but i don't know if that's currently true. http://doc/1iYbKgW2syEguKfV6bCWFtmpWncpJsFZIGRm7vDICyQY touches on this.

Would the solution to this be similar to SoftwareDeps in Tast? And getting that information similar to the GAIA problem above where info is passed into Tast by Autotest? Can discuss in tonight's meeting.

Patch-set: 8