platform: fix tast platform.MLServiceBootstrap
The Tast on-device test platform.MLServiceBootstrap assumes that
ml-service will be in running state immediatelly, but actually it may take some
time for gdbus to start, making it in post-start state. So
platform.MLServiceBootstrap may fail sometime.
In this CL, I replace the upstart.JobStatus with
upstart.WaitForJobStatus, which will wait until ml-service is in the expected
state, or exceeds the timeout and throws an error.

BUG=chromium:940156
TEST=Tested by Tast on-device test

Change-Id: Iab61ce410322eaf517e4801539755711f5c24dd9
Reviewed-on: https://chromium-review.googlesource.com/1522606
Commit-Ready: Xinglong Luan <alanlxl@chromium.org>
Tested-by: Xinglong Luan <alanlxl@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>
Reviewed-by: Andrew Moylan <amoylan@chromium.org>