tast-tests: Crostini: remove component dir in unmount
Mounting the Crostini component involves making a mount point directory
and mounting the component onto it.  The cleanup code unmounted the
component, but it did not remove the directory, which would break
Crostini for any other user logging into the same Chromebook after
running the tast tests, since the test component has a higher version
number (99999) but is empty, since it was umounted.

Remove the mount point directory after unmounting the component to put
the system back into a consistent state.

BUG=None
TEST=tast run vm.CrostiniStartEverything, then Terminal as another user

Change-Id: I0c2420ad615e2b952771a73ecbfd5cb2fab56eb6
Signed-off-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1524761
Reviewed-by: Dan Erat <derat@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>