platform.MemoryPressure: add Pressure Stall Info reporting
PSI (Pressure Stall Information) is not available until
kernel 4.20, but it's available as patches for 4.19 and
we can experiment with it.

At this point there is no expectation that we'll ever
need to use the PSI values internally in the test, so
instead of parsing /proc/psi/memory, we simply read its
content and log it.

BUG=chromium:934937
TEST=ran the test on 4.19 with PSI, checked numbers.  Also ran on 3.14.

Change-Id: If3a6581d3086ec1e8611b9f94fc1c4394b97d524