tast-tests: Ignore disk wait in security.SandboxedServices.
Make the security.SandboxedServices test ignore errors when
getting info about processes in the 'D' state. It appears
that /proc/<pid>/mounts is sometimes missing while a process
is in this state.

Also add rialto_modem_watchdog to the baseline, as it
appears to sometimes run on veyron_rialto DUTs in the lab,
and make getProcSandboxInfo return partially-read info so we
can log it to aid in debugging.

BUG=chromium:936703
TEST=manual: ran it against a few DUTs

Change-Id: Ic3562421c0422dfc9df31168e354785aa9b6ba19