tast-tests: Make security.ExecStack ignore missing stacks.
Make security.ExecStack avoid reporting an error if it
doesn't find a stack for a process. In the lab, we encounter
this seemingly-impossible scenario once every few days. It
seems likely that we're examining a process while it's in
the process of being torn down.

BUG=chromium:877733
TEST=ran it against caroline

Change-Id: I77298f89966378e6ae8192854d84ebcca95c304d
Reviewed-on: https://chromium-review.googlesource.com/1526177
Commit-Ready: Dan Erat <derat@chromium.org>
Tested-by: Dan Erat <derat@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>