platform.MemoryPressure: add WPR record mode
Setting recordPageSet to true and running the test records
a new WPR archive.

This has the limitation that it needs an existing archive
to run, because the data files are unconditionally downloaded.
However, it is still helpful for refreshing the archive, and
the setup may be rigged to work by creating a fake archive,
and it's no worse than before.

A proper solution (in the works) involves separating out most
of the test into a subpackage, and creating an additional "test"
(with disabled=true) to be used exclusively for creating the
archive.

BUG=none
TEST=ran the test after setting recordPageSet = true

Change-Id: I163f2239287626e90f79876734fdfc0e83508db4
Reviewed-on: https://chromium-review.googlesource.com/1529316
Commit-Ready: Luigi Semenzato <semenzato@chromium.org>
Tested-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>