platform.MemoryPressure: add WPR recorder as new test
Move most of the platform.MemoryPressure code to a separate
mempressure module and add a separate "test" for recording
a WPR page set, both for recording new page sets as our needs
evolve, and for the benefit of external users who do not have
access to the WPR archive in Google Storage.

There is no change in functionality in platform.MemoryPressure,
and the test platform.MemoryPressureRecorder is added.

BUG=none
TEST=ran the tests

Change-Id: Icca50f5580e2b2575cbe21a7587e958bb9e7012e
Reviewed-on: https://chromium-review.googlesource.com/1529318
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>