tast: Use en-US for VirtualKeyboardSuggestions.
Previously this test failed on devices with a Japanese keyboard,
where the input method defaults to "Japanese Keyboard". This
keyboard does not have suggestions so the test failed.

We change the test to always switch to the en-US input method so
that this test can run on any devices. Future plan is to test
all the supported Chrome OS languages in this test.

BUG=chromium:930775
TEST=veyron_minnie
CQ-DEPEND=CL:1496186

Change-Id: Ife6cda5ae2a4c21ec7f97e936edc682ec3879591