Update patch set 16
Patch Set 16:

Ah, you're on travel now. So let me ask a question here instead.

CL:1469064 changed the structure of a11y tree in Chrome side and it would have broken arc.AccessibilityTest. While it's not a big problem for now (since the test is still informational), we want all Tast tests to be promoted to important eventually, and then the story is different. Since we run important tests in Chrome PFQ, such Chrome change would break Chrome PFQ. Furthermore, we're planning to run more Tast tests in Chrome CQ, so this test may block the Chrome CL from landing from the first place. Note that, since we can't submit Chrome CL and Chrome OS CL atomically, this test may make it very difficult to make a11y Chrome changes.

So in short, we have to make sure the test is robust enough to:
- allow cosmetic changes to a11y tree without changing test code, and
- allow matching differently shaped trees even for non-cosmetic changes.

For example, is it sufficient to check that an element A is a descendant of another element B for a list of (A, B)? Such check will make the test robust against tree mutations, such as inserting a container element in a middle layer.

Patch-set: 16
Reviewer: Gerrit User 1118346 <1118346@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>