tast-tests: Return error if component mount failed
When ComponentUpdater requests a LoadComponent with Mount, the service
may return a response that has an empty path (but no error).  Rather
than making each caller consider this case, check for it explicitly in
the ComponentUpdater client and convert it into an error.

This still allows LoadComponent calls with NoMount to return an empty
path, since the path is not used in that case.

BUG=chromium:944139
TEST=tast run vm.CrostiniStartBasic
TEST=tast run printer.AddEpsonPrinter

Change-Id: I7df98b245c6bf7082d53c31b37d49980b3252c10
Signed-off-by: Daniel Verkamp <dverkamp@chromium.org>