platform.MemoryPressure: add a shorter timeout for tab activations
Right now if a chrome renderer crashes and we try to activate it, the
test will wait until the 30 minute timeout hits before stopping with a
failure.  This sets a short 30 second timeout for activations to
avoid waiting so long to fail the test.

This also makes activation failures fatal because they are not
expected behavior.

BUG=none
TEST=ran the test

Change-Id: Ie7c0df0b5d354dff329953aaf7e252427abc5761
Reviewed-on: https://chromium-review.googlesource.com/1529955
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>