Update patch set 2
Patch Set 2:

(6 comments)

i'm uneasy with the amount of duplication between this file and the input file that it was based on. is there any way to share more of the logic? maybe we could pass a second shared file to the "go run" commands executed by go:generate?

also, please include the generated file in this comment -- it should be checked in.

Patch-set: 2
Reviewer: Gerrit User 1000042 <1000042@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>