tast: fix time query test failure in ui.AssistantTextQueries.
We observed an intermittently time query test failure with an
error msg like: "Response time 2019-03-22 22:48:00 +0000 UTC
not within 1m0s of current time 2019-03-22 22:49:00.471541266
+0000 UTC". This can be caused by the error when we interpret
the assistant time (because the reponse only contains hour and
minute) based on the now time, where we copy the date and set
sec and nsec to be 0. For example, when:

real assistant time:  2019-03-22 22:48:59.654321
our interpretation: 2019-03-22 22:48:00.000000
now time: 2019-03-22 22:49:00.123456

The test will fail because the time error is greater than
our 1 minute tolerance.

This change fixes this by cutting off the now time information,
truncating the sec and nsecs parts to keep consistent with the
assistant time format and decrease the error of intepretation.

BUG=None
TEST=local compile and manully test.

Change-Id: Ic382e25179c139849bd552cbeac06e9ed7838de1