platform.MemoryPressure: 10-second averages
With this, a kernelmeter instance tracks and reports vmstat
counter rates over the last 10 seconds (moving average).

Coding note: there's some tension between the notion of "sample"
(aggregate values from /proc/vmstat) and a circular buffer of
samples, and the history of each separate value (max rate).
If maintain each value separately, we can combine the circular
buffer and max rates for that value.  If we combine the values
into a sample object, then we can add a time stamp to it, but
max rates have to be in a separate array.

BUG=none
TEST=ran the test

Change-Id: Id521e26ddcb8096adde9c0daf7302746b491e467