tast-tests: Ignore disk wait in security.SandboxedServices.
Reintroduce code in security.SandboxedServices to ignore
errors when inspecting processes in the disk wait state. I'd
previously removed this, but the "D" state looks like it's
sometimes correlated with the "mnt" file being missing for
likely-broken Chrome processes in the lab:

chronos  12175 17.9  ...  D<   ... [chrome]
chronos  12227  0.0  ...  Z<   ...  \_ [chrome] <defunct>

Also add the rialto_cellular_autoconnect script to the
baseline, as it's occasionally running as root on
veyron_rialto devices.

BUG=chromium:936703
TEST=ran it

Change-Id: I34c746a5309c78c12a7813980c582272de189d01