updateserver: Change the payload hash and size for the new payload.
There is an ongoing change related to adding packages for DLCs and we
needed to change the static update payload we have in tast-test. So we
need to update the size and hash of the payload in the response.

TODO(*): In the future it might be better to dynamically calculate these
two values so these changes are not needed anymore.

BUG=chromium:908994
TEST=tast run <IP> platform.DlcService
CQ-DEPEND=CL:1531834

Change-Id: Ia4274d37fece7c5d814bde3fc366b6d509219463