tast-tests: Crostini: override demo app titles to work around vsh bug
The root cause of https://crbug.com/942199 is not yet fixed, but we can
work around the symptoms: we are finding the wrong window in the DPI
scaling tests because the window from an earlier test uses the same
title.  To avoid this and to keep the subtests decoupled, specify a
unique window title in VerifyAppFromTerminal as well.

BUG=chromium:942199
TEST=tast run vm.CrostiniStartEverything

Change-Id: Ib559fd05e3dbdad5f48e54337c0309b106b5cbcb
Signed-off-by: Daniel Verkamp <dverkamp@chromium.org>