tast-tests: migrate vm.CrostiniStartBasic to use build artifacts
As part of allowing Crostini tests in the commit queue, we need to ensure that
tests are reproducible. For now, this means that the Termina VM and Crostini
container will be tarred up and included with a DUT's build artifacts.

Add the support code to allow using the guest_images.tar build artifact in
Crostini tests.

BUG=chromium:943241
TEST=vm.CrostiniStartBasic on a nami-release or nami-paladin build
CQ-DEPEND=CL:1539928

Change-Id: I0565ade96f713b22511498b677879edfed815c0e