platform.MemoryPressure: customize size of initial set
On larger systems, the size of the initial tab set must
be considerably larger than on the smaller systems in
order to induce swapping when cycling through the set.
To hold the additional tabs, the preallocation must also
be adjusted.  The calculations become too complex, so
we just use empirical values.  Correspondingly, we no
longer compute the "memory equalizing amount".  Instead
we log various mm-related quantities.

BUG=none
TEST=tested with 2, 4, 8 and 16 GB

Change-Id: Ic7015d3e45b25e97408e69172c5baa018eece179