platform.MemoryPressure: measure LRU tab switches
This changes the test in multiple ways.

First, the test no longer does preallocation.  The test timeout
is increased to 2 hours to support the larger devices (16GB RAM).
Preallocation code is left in for test development.

Second, at the end of phase 1, the test performs one set
of tab switches to a number of LRU tabs (also called "cold tabs")
and measures the times for those.

Third, the test tries to be resilient to attempts to switch to
a discarded tab.  A broken devtools connection means that the
tab has either crashed or discarded.  After we detect the first
discard (via the chrome dbus signal) we assume that a broken
connection is due to the discard.  This can have false negatives
but they should be relatively rare.  Before the first discard,
broken connections cause fatal errors.

Fourth, some additional system measurements (direct and computed)
are added.

BUG=none
TEST=ran the test

Change-Id: I47b1209099ddea1fdb077727d1a2745d888fbf11