platform.MemoryPressure: more precise preallocation
This test uses "preallocation" to quickly use up a portion of RAM and
swap space at the beginning, which would otherwise be filled by opening
more tabs, at the cost of a much slower run time.  We argue that the
critical part of the test (starting from no pressure, increasing to
higher pressure until the first discard is reached) is not affected
by preallocation, since without preallocation the earlier tabs are
used briefly, then quickly ignored for the remainder of the test.

Calculating equivalent preallocation is tricky.  Some of it depends
on kernel parameters that are not directly exported, thus it needs
to be reverse-engineered from other quantities (see
kernelmeter.KernelMinFree()). It also depends on how much swap
is free at discard.

This works reasonably well on 2, 4, 8 and 16 GiB devices
(about 16, 20, 25 and 27 tabs).
It also accounts for idiosyncrasies on some systems, like the
minnie which starts swapping with 1GB free RAM (other systems
wait until free RAM is down to 150 to 300MB).

Note that overall the test still does not scale well to 8 and 16 GiB.
The discard margin is set at 5% of total RAM, so for instance
it's set at 800 MiB on a 16 GiB device.  At that point the
system has not started swapping heavily, and the initial set
of tabs appears to fit in the hot anonymous pages, so the cycling
does not trigger swaps.  A later CL will attempt to address this
problem by increasing the size of the initial set.

This CL also piggy-backs a small change: separating the
concepts of "working set size" and "initial set size",
so they can be set to separate values.

BUG=chromium:941269
TEST=tested on 2, 4, 8 and 16 GB systems (all with default zram)

Change-Id: Ie969c8f0600c9729f040c758d5c9586a109ad0b9